requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
constantly==23.10.4
crispy-bootstrap4==2024.10
crispy-bootstrap5==2023.10
cryptography==44.0.0
daphne==4.1.2
python-decouple
dj-database-url==2.1.0
Django==4.2.16
django-crispy-forms==2.3
djangorestframework==3.15.2
python-dotenv
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
psycopg
pycparser==2.22
pyOpenSSL==25.0.0
redis==5.2.1
selenium
service-identity==24.2.0
setuptools==75.8.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
whitenoise
zope.interface==7.2